+++ FakeServer/Common/ApiSettings.cs
@@ -3,5 +3,7 @@
     public class ApiSettings
     {
         public bool UpsertOnPut { get; set; }
+
+        public bool UseResultObject { get; set; }
     }
 }