+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased]
 * offset and limit slice option
+* Option for query to return JSON object
 
 ### [0.2.0] - 2017-07-31
 * Get Controller's route template from constant value